backend/requirements.txt
flask
flask-cors
gunicorn
tensorflow-cpu==2.13.0
numpy
Pillow